% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{exercise1}
\alias{exercise1}
\title{Data for exercise 1}
\usage{
data("exercise1")
}
\description{
A \code{QFeatures} object called \code{lcb3} that has been generated to run
exercise 1. It was generated from the \code{v02-scp.Rmd} vignette that
replicates the SCoPE2 data processing workflow for batch LCB3 up to PSM filtering
based on the SCR. This dataset can be used to demonstrate the use
of \code{computeSCR}.

\code{lcb3} contains 24 assays, each assay contains PSM data for one of
the 24 MS runs
}
\seealso{
Other exercise datasets are available: \link{exercise2},
\link{exercise3}, \link{exercise4}, \link{exercise5}.
}
