% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{exercise2}
\alias{exercise2}
\title{Data for exercise 2}
\usage{
data("exercise2")
}
\description{
The data have been generated to run exercise 2. It was generated
from the \code{v02-scp.Rmd} vignette that replicates the SCoPE2 data
processing workflow for batch LCB3 up to the aggregation of PSMs
to peptides. This dataset can be used to demonstrate the use of
\code{aggregateFeaturesOverAssays}.

It loads 3 objects:
\itemize{
\item \code{lcb3}: contains 24 assays, each assay contains PSM data for one of
the 24 MS runs.
\item \code{peptideAssays}: a vector containing new names for the aggregated
assays
\item \code{remove.duplicates}: a function that can be used as a custom
aggregation function.
}
}
\seealso{
Other exercise datasets are available: \link{exercise1},
\link{exercise3}, \link{exercise4}, \link{exercise5}.
}
