% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{exercise3}
\alias{exercise3}
\title{Data for exercise 3}
\usage{
data("exercise3")
}
\description{
A \code{QFeatures} object called \code{lcb3} that has been generated to run
exercise 3. It was generated from the \code{v02-scp.Rmd} vignette that
replicates the SCoPE2 data processing workflow for batch LCB3 up to
the QC and filtering of single-cells based on the CV. This dataset
can be used to demonstrate the use of \code{subsetByColData}.

\code{lcb3} contains 49 assays:
\itemize{
\item Assay 1-24: PSM data for 24 MS runs
\item Assay 25-48: aggregated peptide data for 24 MS runs
\item Assay 49: joined peptide data containing the sample of all 24
MS runs
}
}
\seealso{
Other exercise datasets are available: \link{exercise1},
\link{exercise2}, \link{exercise4}, \link{exercise5}.
}
