% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{exercise4}
\alias{exercise4}
\title{Data for exercise 4}
\usage{
data("exercise4")
}
\description{
The data have been generated to run exercise 4. It was generated
from the \code{v02-scp.Rmd} vignette that replicates the SCoPE2 data
processing workflow for batch LCB3 up to batch correction. This
dataset can be used to demonstrate how to add
\code{SingleCellExperiment} objects as new assay in a \code{QFeatures}
object and create links with a parent assay.

\code{lcb3} contains 55 assays:
\itemize{
\item Assay 1-24: PSM data for 24 MS runs
\item Assay 25-48: aggregated peptide data for 24 MS runs
\item Assay 49: joined peptide data containing the sample of all 24
MS runs
\item Assay 50: peptide data after a first normalization
\item Assay 51: peptide data after a second normalization
\item Assay 52: peptide data after log-transformation
\item Assay 53: aggregated protein data
\item Assay 54: protein data after a first normalization
\item Assay 55: protein data after a second normalization and imputation
of missing data
}

\code{sce} corresponds to Assay 55 after batch correction using \code{ComBat}.
}
\seealso{
Other exercise datasets are available: \link{exercise1},
\link{exercise2}, \link{exercise3}, \link{exercise5}.
}
