% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcb3_1}
\alias{lcb3_1}
\title{Data for exercise 1}
\format{
An object of class \code{QFeatures} of length 24.
}
\usage{
data("lcb3_1")
}
\description{
A \code{QFeatures} object that has been generated to run
exercise 1. It was generated from the \code{v02-scp.Rmd} vignette that
replicates the SCoPE2 data processing workflow for batch LCB3 up to PSM filtering
based on the SCR. This dataset can be used to demonstrate the use
of \code{computeSCR}.

\code{lcb3_1} contains 24 assays, each assay contains PSM data for one of
the 24 MS runs
}
\seealso{
Other exercise datasets are available: \link{lcb3_2},
\link{lcb3_3}, \link{lcb3_4}, \link{lcb3_5}.
}
\keyword{datasets}
