% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcb3_3}
\alias{lcb3_3}
\title{Data for exercise 3}
\format{
An object of class \code{QFeatures} of length 49.
}
\usage{
data("lcb3_3")
}
\description{
A \code{QFeatures} object that has been generated to run
exercise 3. It was generated from the \code{v02-scp.Rmd} vignette that
replicates the SCoPE2 data processing workflow for batch LCB3 up to
the QC and filtering of single-cells based on the CV. This dataset
can be used to demonstrate the use of \code{subsetByColData}.

\code{lcb3_3} contains 49 assays:
\itemize{
\item Assay 1-24: PSM data for 24 MS runs
\item Assay 25-48: aggregated peptide data for 24 MS runs
\item Assay 49: joined peptide data containing the sample of all 24
MS runs
}
}
\seealso{
Other exercise datasets are available: \link{lcb3_1},
\link{lcb3_2}, \link{lcb3_4}, \link{lcb3_5}.
}
\keyword{datasets}
