% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcb3_4}
\alias{lcb3_4}
\title{Data for exercise 4}
\format{
An object of class \code{QFeatures} of length 55.
}
\usage{
data("lcb3_4")
}
\description{
A \code{QFeatures} object that has been generated to run exercise 4. It was generated
from the \code{v02-scp.Rmd} vignette that replicates the SCoPE2 data
processing workflow for batch LCB3 up to batch correction. This
dataset can be used to demonstrate how to add
\code{SingleCellExperiment} objects as new assay in a \code{QFeatures}
object and create links with a parent assay.

\code{lcb3_4} contains 55 assays:
\itemize{
\item Assay 1-24: PSM data for 24 MS runs
\item Assay 25-48: aggregated peptide data for 24 MS runs
\item Assay 49: joined peptide data containing the sample of all 24
MS runs
\item Assay 50: peptide data after a first normalization
\item Assay 51: peptide data after a second normalization
\item Assay 52: peptide data after log-transformation
\item Assay 53: aggregated protein data
\item Assay 54: protein data after a first normalization
\item Assay 55: protein data after a second normalization and imputation
of missing data
}

\code{sce} corresponds to Assay 55 after batch correction using \code{ComBat}.
}
\seealso{
Other exercise datasets are available: \link{lcb3_1},
\link{lcb3_2}, \link{lcb3_3}, \link{lcb3_5}.
}
\keyword{datasets}
