% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcb3_5}
\alias{lcb3_5}
\title{Data for exercise 5}
\format{
An object of class \code{QFeatures} of length 58.
}
\usage{
data("lcb3_5")
}
\description{
A \code{QFeatures} object that has been generated to run
exercise 5. It was generated from the \code{v02-scp.Rmd} vignette that
replicates the SCoPE2 data processing workflow for batch LCB3 up to
data visualization. This dataset can be used to demonstrate the use
of \code{longFormat}.

\code{lcb3_5} contains 58 assays:
\itemize{
\item Assay 1-24: PSM data for 24 MS runs
\item Assay 25-48: aggregated peptide data for 24 MS runs
\item Assay 49: joined peptide data containing the sample of all 24
MS runs
\item Assay 50: peptide data after a first normalization
\item Assay 51: peptide data after a second normalization
\item Assay 52: peptide data after log-transformation
\item Assay 53: aggregated protein data
\item Assay 54: protein data after a first normalization
\item Assay 55: protein data after a second normalization and imputation
of missing data
\item Assay 56: batch corrected protein data
\item Assay 57: batch corrected protein data after a first normalization
\item Assay 58: batch corrected protein data after a second normalization
}
}
\seealso{
Other exercise datasets are available: \link{lcb3_1},
\link{lcb3_2}, \link{lcb3_3}, \link{lcb3_4}.
}
\keyword{datasets}
